feat: `request` can now request data
Right now the API is Thenable so it can be used with `await` easily.

It also supports a `subscribe` method that just pushes the data into
the specified callback but should later be able to pump through progress
updates and maybe errors? The `Observable` API also provides `onError`
and `onComplete` callbacks, this might be overkill for our use-case.